Organized the authentication files into a folder
