+++ src/Authentication/LoggedinIcon.js
